Refactor to kill off shelving.query
The `shelving.query` namespace was a historical accident of the fact that first I designed an API
and then wrote a query compiler against it. Queries are an essential part of the platform I want to
provide, and furthermore the query system is deeply intertwined with shelving's code loading
generally. The query system never really worked as a separate module because of circular
dependencies through `shelving.impl`, so just kill it off by merging the query namespace as required
into `shelving.impl`.

Fixes #26 into the bargan